requirements.txt
fastapi
uvicorn
newspaper3k
pydantic
lxml[html_clean]
motor
pymongo[srv]
python-dotenv